requirements.txt
customtkinter
pillow
colorama
requests
psutil
python-dotenv
pywin32; platform_system == "Windows"
